refactor of main container component
