+++ src/components/Modal/index.css
@@ -6,7 +6,7 @@
     z-index: 700;
     outline: 0;
     width: 60%;
-    height: 60%;
+    height: 80%;
 }
 
 .modal-backdrop {
