--- src/containers/Background/index.css
@@ -1,9 +1,0 @@
-.background {
-    height: 100vh;
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-    z-index: 1;
-}
-
-
