+++ src/containers/Container/index.css
@@ -1,0 +1,9 @@
+.background {
+    height: 100vh;
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+    z-index: 1;
+}
+
+
